# apps/wallet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "relational-rust-server",
    "description": "Custodial Avalanche wallet service running inside Intel SGX enclave with Gramine.",
    "contact": {
      "name": "Relational Network"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/bookmarks": {
      "get": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "List bookmarks for a wallet.",
        "description": "Returns all bookmarks for a wallet owned by the authenticated user.",
        "operationId": "list_bookmarks",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WalletAddress"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bookmark"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your wallet"
          },
          "404": {
            "description": "Wallet not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Create a new bookmark.",
        "description": "Creates a bookmark in a wallet owned by the authenticated user.",
        "operationId": "create_bookmark",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bookmark"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your wallet"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/bookmarks/{bookmark_id}": {
      "delete": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Delete a bookmark.",
        "description": "Deletes a bookmark owned by the authenticated user.",
        "operationId": "delete_bookmark",
        "parameters": [
          {
            "name": "bookmark_id",
            "in": "path",
            "description": "Identifier of the bookmark to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bookmark deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your bookmark"
          },
          "404": {
            "description": "Bookmark not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/invite": {
      "get": {
        "tags": [
          "Invites"
        ],
        "operationId": "get_invite",
        "parameters": [
          {
            "name": "invite_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invite/redeem": {
      "post": {
        "tags": [
          "Invites"
        ],
        "operationId": "redeem_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Bookmark": {
        "type": "object",
        "description": "A saved wallet address bookmark.\n\nBookmarks allow users to save frequently-used addresses with friendly names\nfor quick access when sending transactions.",
        "required": [
          "id",
          "wallet_id",
          "name",
          "address"
        ],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The bookmarked wallet address."
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this bookmark."
          },
          "name": {
            "type": "string",
            "description": "User-friendly name for the bookmarked address."
          },
          "wallet_id": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The wallet this bookmark belongs to."
          }
        }
      },
      "CreateBookmarkRequest": {
        "type": "object",
        "description": "Request to create a new bookmark.",
        "required": [
          "wallet_id",
          "name",
          "address"
        ],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The wallet address to bookmark."
          },
          "name": {
            "type": "string",
            "description": "User-friendly name for the bookmark."
          },
          "wallet_id": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The wallet to add the bookmark to (must be owned by the user)."
          }
        }
      },
...
      "Invite": {
        "type": "object",
        "description": "An invitation code for new users.\n\nInvites can be used to control access to the wallet service. Each invite\ncode can only be redeemed once.",
        "required": [
          "id",
          "code",
          "redeemed"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The invite code (shared with the invitee)."
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this invite."
          },
          "redeemed": {
            "type": "boolean",
            "description": "Whether this invite has been used."
          }
        }
      },
...
      "RedeemInviteRequest": {
        "type": "object",
        "description": "Request to redeem an invite code.",
        "required": [
          "invite_id"
        ],
        "properties": {
          "invite_id": {
            "type": "string",
            "description": "The invite ID to redeem."
          }
        }
      },
...
      "WalletAddress": {
        "type": "string",
        "description": "Ethereum-compatible wallet address wrapper.\n\nProvides type safety for wallet addresses throughout the API.\nFormat: `0x` followed by 40 hexadecimal characters (20 bytes).\n\n# Example\n\n```rust,ignore\nlet addr = WalletAddress::from(\"0x742d35Cc6634C0532925a3b844Bc9e7595f4aB12\");\n```"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk JWT token"
      }
    }
  },
...
}